requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3
langchain>=0.1.6
langchain-community>=0.0.20
langchain-google-genai>=1.0.1
langchain-core>=0.1.23
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
pydantic>=2.6.0
google-generativeai>=0.4.0
sentence-transformers>=2.3.1
python-dotenv>=1.0.1
pandas>=2.0.0
langchain-experimental>=0.0.50